refactor: use OpenAPI to generate API code
