+++ .gitignore
@@ -41,3 +41,6 @@ build
 
 # environment variables
 .env
+
+# swagger example, for learning only
+src/services/swagger/
