+++ config/config.ts
@@ -58,10 +58,9 @@ export default defineConfig({
   openAPI: [
     {
       requestLibPath: "import { request } from 'umi'",
-      // 或者使用在线的版本
-      // schemaPath: "https://gw.alipayobjects.com/os/antfincdn/M%24jrzTTYJN/oneapi.json"
       schemaPath: join(__dirname, 'oneapi.json'),
       mock: false,
+      projectName: 'nanores-cloud',
     },
     {
       requestLibPath: "import { request } from 'umi'",
