+++ mock/task.ts
@@ -7,7 +7,7 @@ const tasks = [
     task_description: 'Mapping the Distribution of Mitochondrial Proteins',
     create_time: new Date('2023-08-20T09:30:51.01'),
     update_time: new Date('2023-08-23T08:25:15.26'),
-    owner_id: 1,
+    user_id: '1',
     input_file_id: 2,
     task_config_id: 2,
     total_cycles: 60,
@@ -20,7 +20,7 @@ const tasks = [
     task_description: 'Synaptic Vesicle Proteins in Neurons',
     create_time: new Date('2023-07-15T11:45:23.17'),
     update_time: new Date('2023-09-03T07:57:09.35'),
-    owner_id: 1,
+    user_id: '1',
     input_file_id: 3,
     task_config_id: 2,
     total_cycles: 60,
@@ -33,7 +33,7 @@ const tasks = [
     task_description: 'Membrane Protein Clustering in Bacteria',
     create_time: new Date('2023-10-02T16:37:52.19'),
     update_time: new Date('2023-10-03T17:45:32.37'),
-    owner_id: 1,
+    user_id: '1',
     input_file_id: 5,
     task_config_id: 2,
     total_cycles: 60,
@@ -46,7 +46,7 @@ const tasks = [
     task_description: 'draft',
     create_time: new Date('2023-10-05T16:37:52.19'),
     update_time: new Date('2023-10-06T17:45:32.37'),
-    owner_id: 1,
+    user_id: '1',
     input_file_id: 5,
     task_config_id: 2,
     total_cycles: 60,
@@ -59,7 +59,7 @@ const tasks = [
     task_description: 'Mapping the Distribution of Mitochondrial Proteins',
     create_time: new Date('2023-06-15T08:52:01.03'),
     update_time: new Date('2023-08-31T21:05:50.23'),
-    owner_id: 2,
+    user_id: '2',
     input_file_id: 12,
     task_config_id: 9,
     total_cycles: 60,
@@ -69,12 +69,15 @@ const tasks = [
 ];
 
 export default {
-  'GET /api/smlm-task-2d/getAllTasks': (req: Request, res: Response, u: string) => {
-    res.json({ data: tasks });
-  },
-  'POST /api/smlm-task-2d/getUserTasks': (req: Request, res: Response, u: string) => {
-    const { owner_id } = req.body;
-    const userTasks = tasks.filter((task) => task.owner_id === owner_id);
-    res.send({ data: userTasks });
+  'GET /api/task': (req: Request, res: Response, u: string) => {
+    const { user_id } = req.query;
+    const userTasks = tasks.filter((task) => task.user_id === user_id);
+    const result = {
+      data: userTasks,
+      total: userTasks.length,
+      success: true,
+    };
+
+    return res.json(result);
   },
 };
