+++ src/app.tsx
@@ -6,7 +6,7 @@ import { PageLoading, SettingDrawer } from '@ant-design/pro-components';
 import type { RunTimeLayoutConfig } from 'umi';
 import { history, Link } from 'umi';
 import defaultSettings from '../config/defaultSettings';
-import { currentUser as queryCurrentUser } from './services/api/api';
+import { currentUser as queryCurrentUser } from './services/nanores-cloud/api';
 
 const isDev = process.env.NODE_ENV === 'development';
 const loginPath = '/user/login';
