+++ src/pages/File.tsx
@@ -1,4 +1,4 @@
-import { addRule, removeRule, rule, updateRule } from '@/services/api/api';
+import { addRule, removeRule, rule, updateRule } from '@/services/nanores-cloud/rule';
 import { PlusOutlined } from '@ant-design/icons';
 import type { ActionType, ProColumns, ProDescriptionsItemProps } from '@ant-design/pro-components';
 import {
