+++ src/services/api/auth.ts
@@ -4,21 +4,21 @@ import { request } from 'umi';
 
 /* send authentication to backend */
 export async function authVerify(
+  params: {
+    access_token: string;
+    expires_at: string;
+    expires_in: string;
+    refresh_token: string;
+    token_type: string;
+    type: string;
+  },
+  options?: { [key: string]: any },
+) {
+  return request('/v1/user/verify', {
+    method: 'GET',
     params: {
-      access_token: string;
-      expires_at: string;
-      expires_in: string;
-      refresh_token: string;
-      token_type: string;
-      type: string;
+      ...params,
     },
-    options?: { [key: string]: any },
-  ) {
-    return request('/v1/user/verify', {
-      method: 'GET',
-      params: {
-        ...params,
-      },
-      ...(options || {}),
-    });
-  }
+    ...(options || {}),
+  });
+}
