+++ src/services/nanores-cloud/api.ts
@@ -1,0 +1,11 @@
+// @ts-ignore
+/* eslint-disable */
+import { request } from 'umi';
+
+/** 获取当前的用户 GET /api/currentUser */
+export async function currentUser(options?: { [key: string]: any }) {
+  return request<API.CurrentUser>('/api/currentUser', {
+    method: 'GET',
+    ...(options || {}),
+  });
+}
