+++ src/services/nanores-cloud/index.ts
@@ -2,9 +2,15 @@
 /* eslint-disable */
 // API 更新时间：
 // API 唯一标识：
+import * as task from './task';
+import * as config from './config';
 import * as api from './api';
 import * as login from './login';
+import * as rule from './rule';
 export default {
+  task,
+  config,
   api,
   login,
+  rule,
 };
