--- src/services/swagger/index.ts
@@ -1,12 +1,0 @@
-// @ts-ignore
-/* eslint-disable */
-// API 更新时间：
-// API 唯一标识：
-import * as pet from './pet';
-import * as store from './store';
-import * as user from './user';
-export default {
-  pet,
-  store,
-  user,
-};
