--- src/services/swagger/user.ts
@@ -1,114 +1,0 @@
-// @ts-ignore
-/* eslint-disable */
-import { request } from 'umi';
-
-/** Create user This can only be done by the logged in user. POST /user */
-export async function createUser(body: API.User, options?: { [key: string]: any }) {
-  return request<any>('/user', {
-    method: 'POST',
-    data: body,
-    ...(options || {}),
-  });
-}
-
-/** Creates list of users with given input array POST /user/createWithArray */
-export async function createUsersWithArrayInput(
-  body: API.User[],
-  options?: { [key: string]: any },
-) {
-  return request<any>('/user/createWithArray', {
-    method: 'POST',
-    data: body,
-    ...(options || {}),
-  });
-}
-
-/** Creates list of users with given input array POST /user/createWithList */
-export async function createUsersWithListInput(body: API.User[], options?: { [key: string]: any }) {
-  return request<any>('/user/createWithList', {
-    method: 'POST',
-    data: body,
-    ...(options || {}),
-  });
-}
-
-/** Logs user into the system GET /user/login */
-export async function loginUser(
-  params: {
-    // query
-    /** The user name for login */
-    username: string;
-    /** The password for login in clear text */
-    password: string;
-  },
-  options?: { [key: string]: any },
-) {
-  return request<string>('/user/login', {
-    method: 'GET',
-    params: {
-      ...params,
-    },
-    ...(options || {}),
-  });
-}
-
-/** Logs out current logged in user session GET /user/logout */
-export async function logoutUser(options?: { [key: string]: any }) {
-  return request<any>('/user/logout', {
-    method: 'GET',
-    ...(options || {}),
-  });
-}
-
-/** Get user by user name GET /user/${param0} */
-export async function getUserByName(
-  params: {
-    // path
-    /** The name that needs to be fetched. Use user1 for testing.  */
-    username: string;
-  },
-  options?: { [key: string]: any },
-) {
-  const { username: param0 } = params;
-  return request<API.User>(`/user/${param0}`, {
-    method: 'GET',
-    params: { ...params },
-    ...(options || {}),
-  });
-}
-
-/** Updated user This can only be done by the logged in user. PUT /user/${param0} */
-export async function updateUser(
-  params: {
-    // path
-    /** name that need to be updated */
-    username: string;
-  },
-  body: API.User,
-  options?: { [key: string]: any },
-) {
-  const { username: param0 } = params;
-  return request<any>(`/user/${param0}`, {
-    method: 'PUT',
-    params: { ...params },
-    data: body,
-    ...(options || {}),
-  });
-}
-
-/** Delete user This can only be done by the logged in user. DELETE /user/${param0} */
-export async function deleteUser(
-  params: {
-    // path
-    /** The name that needs to be deleted */
-    username: string;
-  },
-  options?: { [key: string]: any },
-) {
-  const { username: param0 } = params;
-  return request<any>(`/user/${param0}`, {
-    method: 'DELETE',
-    params: { ...params },
-    ...(options || {}),
-  });
-}
